requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
llama-index-core
llama-index-readers-file
llama-index-embeddings-openai
llama-index-vector-stores-faiss
openai
llama-index-llms-openai
langchain